Add some metadata tests and fix issue when the player is deallocated
