+++ Tests/AnalyticsTests/CommandersAct/CommandersActTracker.swift
@@ -13,6 +13,6 @@ extension CommandersActTracker.Metadata {
     }
 
     static func test(streamType: StreamType) -> Self {
-        .init(labels: ["meta": "data"], streamType: streamType)
+        .init(labels: ["media_title": "title"], streamType: streamType)
     }
 }
